requirements.txt
flask
flask-cors
nltk
requests
bs4
python-dotenv
